requirements.txt
scikit-image==0.18.1
PyYAML==5.4.1
tqdm==4.49.0
scikit-learn==0.24.0
pytorch-lightning==1.4.9
opencv-python==4.5.1.48
hydra-core==1.0.5
numpy==1.20.3
numba==0.53
imageio
protobuf==3.19.6
tabulate
kornia==0.5.0
torchmetrics==0.6.0
efficientnet_pytorch==0.7.1
einops==0.6.0
open3d==0.15.2
trimesh
openmim
mmcv-full==1.5.0
mmdet==2.20.0
pyqt5==5.12.0